// src/modules/log-activities/enum/log-action.enum.ts
export enum LogAction {
  MARK_READ = 'mark_read',
  MARK_UNREAD = 'mark_unread',
  MARK_NEW = 'mark_new',
  MARK_EXISTING = 'mark_existing',
  MARK_COMPLETED = 'mark_completed',
  MARK_INCOMPLETE = 'mark_incomplete',
  SELECT_LABELS = 'select_labels',
  UPDATE_CLIENT = 'update_client',
  UNDEFINED = 'undefined',
}